Setup script needs to be a bit different between versions.

+++ util/scripts/setup-robot.sh
@@ -1,7 +1,7 @@
 # To be run on the robot AS ROOT
 
-if [ $# -ne 1 ] ; then
-    echo "Usage: ./setup-robot <new-hostname>"
+if [ $# -ne 2 ] ; then
+    echo "Usage: ./setup-robot <new-hostname> <old-or-new>"
     exit 1
 fi
 
@@ -10,13 +10,20 @@ if [ x$(whoami) != xroot ]; then
     exit 1
 fi
 
+if [ $2 != 'old' -a $2 != 'new' ]; then
+    echo "The second parameter needs to be 'old' or 'new'"
+    exit 1
+fi
+
 HOSTNAME=$1
+TYPE=$2
 
 # Make nbites folders
 echo "Making folders for nbites content..."
 mkdir -p nbites/audio nbites/frames nbites/log
 chown -R nao nbites
 rm -rf behaviors
+rm -rf recordings
 
 # Put audio files in their place
 echo "Moving audio files..."
@@ -26,7 +33,11 @@ mv *.wav nbites/audio
 echo "Moving libraries..."
 mv libboost_python-mt.so /usr/lib/
 mv libprotobuf.so.7 /usr/lib/
-mv lxv4l2.ko /lib/modules/2.6.29.6-rt24-aldebaran-rt/kernel/drivers/media/video/lxv4l2/
+if [ $TYPE == 'old' ]; then
+    mv lxv4l2.ko /lib/modules/2.6.29.6-rt24-aldebaran-rt/kernel/drivers/media/video/lxv4l2/
+elif [ $TYPE == 'new' ]; then
+    rm lxv4l2.ko
+fi
 
 # Set the hostname
 echo "Setting the hostname to $HOSTNAME..."
@@ -35,7 +46,7 @@ echo $HOSTNAME > /etc/hostname
 # Set up the profile
 echo "Adding nbites config to profile..."
 echo "# NBites Config below:" >> /etc/profile
-echo "alias naolog='tail -f /home/nao/nbites/log/naoqi.log'" >> /etc/profile
+echo "alias naolog='tail -f /home/nao/nbites/log/nblog'" >> /etc/profile
 echo "ulimit -S -c unlimited" >> /etc/profile
 
 # Move the etc config
